Local token example (#2)
* Replace expired certificates

* STS certificate

* Refactor channelstack configuration

Pass the entire STS configuration down the stack instead of just the STS certificate. We are going to need more of the configuration
update

* Dictionary update to remove warnings about spelling

Added "Digst" and "Idws"

* Configuration refactoring

* Update certificate and DNS references in ConfbyCode example

* Refactor channelstack configuration

Pass the entire STS configuration down the stack instead of just the STS certificate. We are going to need more of the configuration
update

* Dictionary update to remove warnings about spelling

Added "Digst" and "Idws"

* Configuration refactoring

* Certificate and Modify auto-generated proxy to use correct protection levels

plus Color coded setup script

* Working example of local token

* Code comments, code cleanup, configuration settings

* Documentation

* Move classes into files. Documentation.

* Upgraded all projects to .Net 4.7.2
Updated WSC and STS certificate on all projects
Fixed IIS ekspres issue in WscBootstrapExample.

* Updated WSP certificate and DNS name in examples and integrations tests.

Co-authored-by: Uffe Seerup <use@schultz.dk>
Co-authored-by: Kasper Vestergaard Møller <kvm@schultz.dk>
Co-authored-by: Kasper Vestergaard Møller <mollekas@gmail.com>